refactor: move home page to features
