+++ site/src/features/pages/home/index.ts
@@ -1,0 +1,3 @@
+export { Home } from './components/page'
+
+export type { HomeProps } from './components/page'
